Update RapidezServiceProvider.php
When a view composer is executed in a ServiceProvider of another package, it does not always go well. An example is https://github.com/rapidez/amasty-promo-banners/blob/master/src/AmastyPromoBannersServiceProvider.php#L65. Because this starts alphabetically earlier, it is executed before the Themes are booted. Moving it to a registry fixes it.